Core modules order and migrations fixes

+++ src/CoreModule/AuthorizationModule/migrations/20201207203724_create_roles_actions.php
@@ -23,7 +23,7 @@ final class CreateRolesActions extends AbstractMigration
 		$roles_actions
 			->addColumn('roles_id', 'integer')
 			->addColumn('actions_id', 'integer')
-			->addColumn('authorized', 'smallint')
+			->addColumn('authorized', 'integer')
 			->addForeignKey('roles_id', 'roles', 'id')
 			->addForeignKey('actions_id', 'actions', 'id')
 			->create();
